# other/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Welcome to the nodejs backend of Booboky!"
  version: "1.0.0"
  title: "Booboky"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/api"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /user/register:
    post:
      tags:
      - "user"
      summary: "Registration"
      description: "Register into the bookstore"
      operationId: "userRegisterPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "succesfull registration"
        400:
          description: "User already present"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/login:
    post:
      tags:
      - "user"
      summary: "Login"
      description: "Login with a form"
      operationId: "userLoginPOST"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body"
      responses:
        200:
          description: "succesfull login"
        404:
          description: "unauthorized"
      x-swagger-router-controller: "User"
  /user/logout:
    get:
      tags:
      - "user"
      summary: "Logs out the current user session"
      description: "Logout"
      operationId: "userLogoutGET"
      parameters: []
      responses:
        200:
          description: "Status 200"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
  /user/{id}/address:
    get:
      tags:
      - "user"
      - "address"
      summary: "Get the list of addresses for the user"
      operationId: "userAddressGET"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Address"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
    post:
      tags:
      - "address"
      - "user"
      summary: "Add new address for the user"
      operationId: "userAddressPOST"
      consumes: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Address"
      responses:
        200:
          description: "Status 200"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "Address"
    put:
      tags:
      - "address"
      - "user"
      summary: "Update an exsisting address"
      operationId: "userAddressPUT"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "addressID"
        in: "query"
        description: "The identifier for the address to update"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Address"
      responses:
        200:
          description: "Status 200"
        400:
          description: "Invalid address ID - the address to update does not exists"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "Address"
    delete:
      tags:
      - "user"
      - "address"
      summary: "Delete an existing address for the user"
      operationId: "userAddressDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "addressID"
        in: "query"
        description: "The id of the address to be removed"
        required: true
        type: "integer"
      responses:
        200:
          description: "The address has been removed"
        400:
          description: "The address id is invalid"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
  /user/{id}/whishlist:
    get:
      tags:
      - "user"
      summary: "Get the list of books in the whishlist"
      operationId: "userWhishlistGET"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      summary: "Removes one book from the whishlist"
      operationId: "userWhishlistDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "bookID"
        in: "query"
        description: "The id of the book to be removed from the whishlist"
        required: true
        type: "integer"
      responses:
        200:
          description: "Status 200"
        400:
          description: "Invalid book id - the book is not presenti on the whishlist"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
  /user/{id}/order:
    get:
      tags:
      - "user"
      summary: "Get the orders for a given user"
      operationId: "userOrderGET"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
  /user/{id}/chart:
    get:
      tags:
      - "user"
      - "chart"
      summary: "Get the user's current chart"
      operationId: "userChartGET"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
          schema:
            $ref: "#/definitions/Chart"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "user"
      - "chart"
      summary: "Updates the current chart"
      operationId: "userChartPUT"
      consumes:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body_1"
      responses:
        200:
          description: "Status 200"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "user"
      - "chart"
      summary: "Delete the current chart"
      operationId: "userChartDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Status 200"
        404:
          description: "Status 404"
      security:
      - OAuth2:
        - "admin"
        - "user"
      x-swagger-router-controller: "User"
...
securityDefinitions:
  OAuth2:
    type: "oauth2"
    tokenUrl: "https://localhost/api/auth"
    flow: "password"
    scopes:
      admin: "Website administrator"
      user: "Registred user"
definitions:
  Book:
    type: "object"
    required:
    - "author"
    - "price"
    - "publisher"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      author:
        type: "string"
      price:
        $ref: "#/definitions/Amount"
      image_url:
        type: "array"
        items:
          type: "string"
      genres:
        type: "array"
        items:
          type: "string"
      publisher:
        type: "string"
    description: "A book for sale in the store"
    example: "{\n  id: 0,\n  title: \"Il deserto dei tartari\",\n  author: \"Dino\
      \ Buzzati\",\n  price: {\n    value: 10,\n    currency: \"eur\"\n  },\n  image_url:\
      \ \"https://test.com\"\n}"
  Amount:
    type: "object"
    required:
    - "value"
    properties:
      value:
        type: "number"
        format: "double"
        minimum: 0.01
        maximum: 1000
    description: "Price\n"
    example: "{\n  currency: \"eur\",\n  value: 11\n}"
  User:
    type: "object"
    required:
    - "email"
    - "name"
    - "surname"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
      birthdate:
        type: "string"
        format: "date"
    description: "A user"
    example: "{\n  id: 1,\n  name: \"Alan\",\n  surname: \"Turing\",\n  email: \"\
      alan.turing@example.com\"\n  birthdate: \"1989-01-01\"\n}"
...
  Address:
    type: "object"
    required:
    - "address_line_1"
    - "cap"
    - "city"
    - "country"
    - "name"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      address_line_1:
        type: "string"
      address_line_2:
        type: "string"
      cap:
        type: "integer"
      city:
        type: "string"
      country:
        type: "string"
    description: "An address -.-"
    example:
      country: "country"
      cap: 6
      city: "city"
      name: "name"
      address_line_1: "address_line_1"
      id: 0
      address_line_2: "address_line_2"
...
  Chart:
    type: "object"
    required:
    - "Books"
    - "User"
    - "timestamp"
    - "total_amount"
    properties:
      User:
        $ref: "#/definitions/User"
      Books:
        type: "array"
        description: "The list of books ordered"
        items:
          $ref: "#/definitions/Order_Books"
      total_amount:
        type: "number"
        description: "Amount to pay for this order"
      timestamp:
        type: "string"
        format: "date-time"
    description: "An order for some books"
    example: "{\n  OrderID: 123456789,\n  User: \n  {\n   id: 1,\n   name: \"Alan\"\
      ,\n   surname: \"Turing\",\n   email: \"alan.turing@example.com\"\n   birthdate:\
      \ \"1989-01-01\"\n}\n  Books: [\n  {\n  id: 0,\n  title: \"Il deserto dei tartari\"\
      ,\n  author: \"Dino Buzzati\",\n  price: {\n    value: 10,\n    currency: \"\
      eur\"\n  }\n\n},\n{\n  id: 1,\n  title: \"Il deserto dei tartari\",\n  author:\
      \ \"Dino Buzzati\",\n  price: {\n    value: 10,\n    currency: \"eur\"\n  }\n\
      }\n  ]\n}"
  body:
    type: "object"
    required:
    - "password"
    - "username"
    properties:
      username:
        type: "string"
      password:
        type: "string"
  body_1:
    type: "object"
    required:
    - "bookID"
    - "qty"
    properties:
      bookID:
        type: "integer"
      qty:
        type: "integer"
  Order_Books:
    required:
    - "book"
    - "qty"
    properties:
      book:
        $ref: "#/definitions/Book"
      qty:
        type: "integer"